requirements_keras.txt
ipython==7.11.1
Pillow==7.0.0
matplotlib==3.1.2
tensorflow==1.14.0
keras==2.2.4